(NO MERGE) Test integration of debug logging
This is a feature originally designed by LoungeKatt and implemented in
reicast to retrieve logs when the application crashed. This feature
allows both on-crash and on-demand logging of the primary app
interfaces.

This feature does not support merging / cherry-pick at this time, as it
has not been rewritten from the reicast version.

If tests are successful, a new version derived from the original
LoungeKatt library source will be made available.