Various UI changes.
Update logos (made them bigger and more aligned on xhdpi and xxhdpi),
update place holder text to use strings, use logos instead of title and
subtext, change titles to be "- Games" via the string but in the future
it should change automatically to be a header so the user knows where
they are at such as Recent, Everything, and Favorites. Removed pointless
subtext and title for navigation drawer.